requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
openpyxl==3.1.2
requests==2.32.5
beautifulsoup4==4.13.3
lxml==5.1.0
Werkzeug==3.1.3
